Move web-layer tests to LightNap.WebApi.Tests
Moves AnonymousVisitorIdMiddlewareTests to the dedicated WebApi test
project. Also applies the UserContextTests split that
feature/iusercontext-multi-actor introduces (this branch was based on
that one, so it carried the combined file): SystemUserContextTests
stays in LightNap.Core.Tests while WebUserContextTests moves to
LightNap.WebApi.Tests. Removes the LightNap.WebApi project reference
from LightNap.Core.Tests that was inverting the dependency graph.
