+++ src/LightNap.Core.Tests/LightNap.Core.Tests.csproj
@@ -23,7 +23,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\LightNap.Core\LightNap.Core.csproj" />
-    <ProjectReference Include="..\LightNap.WebApi\LightNap.WebApi.csproj" />
   </ItemGroup>
 
   <ItemGroup>
